refactor: 마이페이지 폴더 분리
