+++ apps/client/src/widgets/mypage/components/comment-preview/comment-preview.tsx
@@ -3,7 +3,7 @@ import { Title } from '@bds/ui';
 import { getTimeAgo } from '@shared/utils/get-time-ago';
 
 import * as styles from './comment-preview.css';
-import { contentText } from './post-preview.css';
+import { contentText } from '@widgets/mypage/components/post-preview/post-preview.css';
 
 interface CommentPreviewProps {
   content?: string;
