+++ apps/client/src/widgets/mypage/components/post-preview/post-preview.tsx
